+++ pelican-core/src/test/kotlin/io/github/matthewjones372/pelican/SpiPackageTest.kt
@@ -24,6 +24,8 @@ class SpiPackageTest {
         "routeIndex",
         "RenderedError",
         "renderError",
+        "ClassifiedError",
+        "classifyError",
         "RequestBodyCodecs",
         "requestBodyCodec",
         "readStrictBody",